improvement: extract SafeTransactionSelector component
